+++ Labs_PIIS/lab6/lab6/Point.h
@@ -5,19 +5,19 @@ using namespace std;
 class Point
 {
 private:
-	vector<float> coordinates;
+	vector<double> coordinates;
 
 public:
 	Point(int size);
-	Point(vector<float> p);
+	Point(vector<double> p);
 	Point();
 	
-	float& operator[](int index);
+	double& operator[](int index);
 
 	Point operator+(const Point& point2);
 	Point operator-(const Point& point2);
-	Point operator*(const float number);
-	Point operator/(const float number);
+	Point operator*(const double number);
+	Point operator/(const double number);
 
 	int size();
 };
